wrapping the api inside the dao objects

+++ mybatis-jaxws-api/src/main/java/com/github/dcendents/mybatis/jaxws/api/Film.java
@@ -11,7 +11,7 @@ public class Film {
 	private String description;
 	private Language language;
 	private int year;
-	private int length;
+	private Short length;
 	private String[] categories;
 	private List<Actor> actors;
